refs #7300 - update samples
