+++ samples/client/petstore/go/go-petstore-withXml/.swagger-codegen/VERSION
@@ -1,1 +1,1 @@
-2.4.0-SNAPSHOT
+2.4.3-SNAPSHOT
