+++ samples/client/petstore/go/go-petstore/.swagger-codegen/VERSION
@@ -1,1 +1,1 @@
-2.3.1
+2.4.3-SNAPSHOT
