--- samples/client/petstore/go/go-petstore/array_of_array_of_number_only.go
@@ -1,15 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type ArrayOfArrayOfNumberOnly struct {
-	ArrayArrayNumber [][]float32 `json:"ArrayArrayNumber,omitempty"`
-}
