--- samples/client/petstore/go/go-petstore/array_test.go
@@ -1,17 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type ArrayTest struct {
-	ArrayOfString []string `json:"array_of_string,omitempty"`
-	ArrayArrayOfInteger [][]int64 `json:"array_array_of_integer,omitempty"`
-	ArrayArrayOfModel [][]ReadOnlyFirst `json:"array_array_of_model,omitempty"`
-}
