--- samples/client/petstore/go/go-petstore/capitalization.go
@@ -1,21 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type Capitalization struct {
-	SmallCamel string `json:"smallCamel,omitempty"`
-	CapitalCamel string `json:"CapitalCamel,omitempty"`
-	SmallSnake string `json:"small_Snake,omitempty"`
-	CapitalSnake string `json:"Capital_Snake,omitempty"`
-	SCAETHFlowPoints string `json:"SCA_ETH_Flow_Points,omitempty"`
-	// Name of the pet 
-	ATT_NAME string `json:"ATT_NAME,omitempty"`
-}
