--- samples/client/petstore/go/go-petstore/docs/DefaultApi.md
@@ -1,36 +1,0 @@
-# \DefaultApi
-
-All URIs are relative to *http://petstore.swagger.io:80/v2*
-
-Method | HTTP request | Description
-------------- | ------------- | -------------
-[**TestBodyWithQueryParams**](DefaultApi.md#TestBodyWithQueryParams) | **Put** /fake/body-with-query-params | 
-
-
-# **TestBodyWithQueryParams**
-> TestBodyWithQueryParams(ctx, body, query)
-
-
-### Required Parameters
-
-Name | Type | Description  | Notes
-------------- | ------------- | ------------- | -------------
- **ctx** | **context.Context** | context for authentication, logging, cancellation, deadlines, tracing, etc.
-  **body** | [**User**](User.md)|  | 
-  **query** | **string**|  | 
-
-### Return type
-
- (empty response body)
-
-### Authorization
-
-No authorization required
-
-### HTTP request headers
-
- - **Content-Type**: application/json
- - **Accept**: Not defined
-
-[[Back to top]](#) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to Model list]](../README.md#documentation-for-models) [[Back to README]](../README.md)
-
