+++ samples/client/petstore/go/go-petstore/docs/PetApi.md
@@ -62,7 +62,7 @@ Optional parameters are passed through a pointer to a DeletePetOpts struct
 Name | Type | Description  | Notes
 ------------- | ------------- | ------------- | -------------
 
- **apiKey** | **optional.**|  | 
+ **apiKey** | **optional.String**|  | 
 
 ### Return type
 
@@ -211,8 +211,8 @@ Optional parameters are passed through a pointer to a UpdatePetWithFormOpts stru
 Name | Type | Description  | Notes
 ------------- | ------------- | ------------- | -------------
 
- **name** | **optional.**| Updated name of the pet | 
- **status** | **optional.**| Updated status of the pet | 
+ **name** | **optional.String**| Updated name of the pet | 
+ **status** | **optional.String**| Updated status of the pet | 
 
 ### Return type
 
@@ -249,7 +249,7 @@ Optional parameters are passed through a pointer to a UploadFileOpts struct
 Name | Type | Description  | Notes
 ------------- | ------------- | ------------- | -------------
 
- **additionalMetadata** | **optional.**| Additional data to pass to server | 
+ **additionalMetadata** | **optional.String**| Additional data to pass to server | 
  **file** | **optional.Interface of *os.File**| file to upload | 
 
 ### Return type
