--- samples/client/petstore/go/go-petstore/enum_class.go
@@ -1,20 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type EnumClass string
-
-// List of EnumClass
-const (
-	ABC_EnumClass EnumClass = "_abc"
-	EFG_EnumClass EnumClass = "-efg"
-	XYZ_EnumClass EnumClass = "(xyz)"
-)
