--- samples/client/petstore/go/go-petstore/outer_composite.go
@@ -1,17 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type OuterComposite struct {
-	MyNumber *OuterNumber `json:"my_number,omitempty"`
-	MyString *OuterString `json:"my_string,omitempty"`
-	MyBoolean *OuterBoolean `json:"my_boolean,omitempty"`
-}
