--- samples/client/petstore/go/go-petstore/user.go
@@ -1,23 +1,0 @@
-/*
- * Swagger Petstore
- *
- * This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\
- *
- * API version: 1.0.0
- * Contact: apiteam@swagger.io
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package petstore
-
-type User struct {
-	Id int64 `json:"id,omitempty"`
-	Username string `json:"username,omitempty"`
-	FirstName string `json:"firstName,omitempty"`
-	LastName string `json:"lastName,omitempty"`
-	Email string `json:"email,omitempty"`
-	Password string `json:"password,omitempty"`
-	Phone string `json:"phone,omitempty"`
-	// User Status
-	UserStatus int32 `json:"userStatus,omitempty"`
-}
